requirements.txt
cmake
opencv-python
face_recognition
requests
playsound
psycopg2-binary
flask
